Frontend - refactor types
