+++ frontend/src/pam/mission/actions/action-control-env.tsx
@@ -4,11 +4,12 @@ import Divider from 'rsuite/Divider'
 import { Stack } from 'rsuite'
 import Text from '../../../ui/text'
 import { formatDateTimeForFrenchHumans } from '../../../dates'
-import { FishAction } from '../../fish-mission-types'
+import { FishAction } from '../../../types/fish-mission-types'
 import ControlsToCompleteTag from '../controls/controls-to-complete-tag'
 import EnvControlForm from '../controls/env-control-form'
-import { Action, ControlType } from '../../mission-types'
-import { EnvActionControl, actionTargetTypeLabels, vehicleTypeLabels } from '../../env-mission-types'
+import { Action } from '../../../types/action-types'
+import { ControlType } from '../../../types/control-types'
+import { EnvActionControl, actionTargetTypeLabels, vehicleTypeLabels } from '../../../types/env-mission-types'
 import { GET_MISSION_BY_ID, MUTATION_ADD_OR_UPDATE_INFRACTION_ENV } from '../queries'
 import { useMutation } from '@apollo/client'
 import { useParams } from 'react-router-dom'
