+++ frontend/src/pam/mission/actions/action-control-fish.tsx
@@ -1,10 +1,11 @@
 import React from 'react'
 import { CoordinatesFormat, CoordinatesInput, DatePicker, Icon, THEME } from '@mtes-mct/monitor-ui'
-import { FishAction, formatMissionActionTypeForHumans } from '../../fish-mission-types'
+import { FishAction, formatMissionActionTypeForHumans } from '../../../types/fish-mission-types'
 import Text from '../../../ui/text'
 import { Stack } from 'rsuite'
 import ControlsToCompleteTag from '../controls/controls-to-complete-tag'
-import { Action, ControlType } from '../../mission-types'
+import { Action } from '../../../types/action-types'
+import { ControlType } from '../../../types/control-types'
 import ControlAdministrativeForm from '../controls/control-administrative-form'
 import ControlNavigationForm from '../controls/control-navigation-form'
 import ControlGensDeMerForm from '../controls/control-gens-de-mer-form'
