+++ frontend/src/pam/mission/actions/action-mapping.ts
@@ -1,7 +1,8 @@
 import { FC } from 'react'
-import { ActionTypeEnum } from '../../env-mission-types'
+import { ActionTypeEnum } from '../../../types/env-mission-types'
 import ActionControlEnv from './action-control-env'
-import { Action, isEnvAction, isFishAction, isNavAction } from '../../mission-types'
+import { Action } from '../../../types/action-types'
+import { isEnvAction, isFishAction, isNavAction } from './utils'
 import ActionControlFish from './action-control-fish'
 import ActionControlNav from './action-control-nav'
 import ActionStatusForm from './action-status-form'
