+++ frontend/src/pam/mission/actions/utils.ts
@@ -1,0 +1,14 @@
+import { Action } from '../../../types/action-types'
+import { MissionSourceEnum } from '../../../types/env-mission-types'
+
+export function isEnvAction(action: Action): boolean {
+  return action !== null && action.source === MissionSourceEnum.MONITORENV
+}
+
+export function isFishAction(action: Action): boolean {
+  return action !== null && action.source === MissionSourceEnum.MONITORFISH
+}
+
+export function isNavAction(action: Action): boolean {
+  return action !== null && action.source === MissionSourceEnum.RAPPORTNAV
+}
