+++ frontend/src/pam/mission/general-info/mission-distance-consumption.tsx
@@ -1,8 +1,6 @@
 import React from 'react'
-import { FlexboxGrid, Panel, Stack } from 'rsuite'
-import { THEME, DateRangePicker, Label, TextInput } from '@mtes-mct/monitor-ui'
-import Text from '../../ui/text'
-import { MissionGeneralInfo } from '../mission-types'
+import { Stack } from 'rsuite'
+import { THEME, Label, TextInput } from '@mtes-mct/monitor-ui'
 import { MUTATION_ADD_OR_UPDATE_DISTANCE_CONSUMPTION } from '../queries'
 import { useMutation } from '@apollo/client'
 import omit from 'lodash/omit'
