+++ frontend/src/pam/mission/infractions/infraction-summary.tsx
@@ -2,7 +2,8 @@ import React from 'react'
 import { Stack } from 'rsuite'
 import { Accent, Icon, Size, THEME, IconButton, Tag } from '@mtes-mct/monitor-ui'
 import Text from '../../../ui/text'
-import { ControlType, Infraction } from '../../mission-types'
+import { ControlType } from '../../../types/control-types'
+import { Infraction } from '../../../types/infraction-types'
 import { infractionTitleForControlType } from './utils'
 import InfractionTag from './infraction-tag'
