+++ frontend/src/types/env-mission-types.ts
