+++ frontend/src/types/fish-mission-types.ts
